requirements.txt
confluent-kafka==2.4.0
redis==5.0.1
pytest==7.4.3
docker==6.1.3
python-dotenv==1.0.0
